requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
Werkzeug==3.0.1
psutil==5.9.5
email-validator==2.1.0
